Optimize particle loop using float reciprocal square root and fast ARM math compilation

+++ flurry-wayland/flurry-star.c
@@ -55,32 +55,34 @@ void InitStar(Star *s)
 
 void UpdateStar(global_info_t *global, flurry_info_t *flurry, Star *s)
 {
-    float rotationsPerSecond = (float) (2.0*PI*12.0/MAXANGLES) * s->rotSpeed /* speed control */;
-    double thisPointInRadians;
-    double thisAngle = flurry->fTime*rotationsPerSecond;
+    float rotationsPerSecond = (2.0f * (float)PI * 12.0f / (float)MAXANGLES) * s->rotSpeed;
+    float thisPointInRadians;
+    float thisAngle = (float)(flurry->fTime * (double)rotationsPerSecond);
     float cf;
-    double tmpX1,tmpY1,tmpZ1;
-    double tmpX2,tmpY2,tmpZ2;
-    double tmpX3,tmpY3,tmpZ3;
-    double tmpX4,tmpY4,tmpZ4;
-    double rotation;
-    double cr;
-    double sr;
+    float tmpX1,tmpY1,tmpZ1;
+    float tmpX2,tmpY2,tmpZ2;
+    float tmpX3,tmpY3,tmpZ3;
+    float tmpX4,tmpY4,tmpZ4;
+    float rotation;
+    float cr;
+    float sr;
 
     s->ate = 0;
     
-    cf = ((float) (cos(7.0*((flurry->fTime)*rotationsPerSecond))+cos(3.0*((flurry->fTime)*rotationsPerSecond))+cos(13.0*((flurry->fTime)*rotationsPerSecond))));
+    cf = (cosf(7.0f * (float)(flurry->fTime * (double)rotationsPerSecond)) +
+          cosf(3.0f * (float)(flurry->fTime * (double)rotationsPerSecond)) +
+          cosf(13.0f * (float)(flurry->fTime * (double)rotationsPerSecond)));
     cf /= 6.0f;
     cf += 0.75f; 
-    thisPointInRadians = 2.0 * PI * (double) s->mystery / (double) BIGMYSTERY;
+    thisPointInRadians = 2.0f * (float)PI * (float)s->mystery / (float)BIGMYSTERY;
     
-    s->position[0] = 250.0f * cf * (float) cos(11.0 * (thisPointInRadians + (3.0*thisAngle)));
-    s->position[1] = 250.0f * cf * (float) sin(12.0 * (thisPointInRadians + (4.0*thisAngle)));
-    s->position[2] = 250.0f * (float) cos((23.0 * (thisPointInRadians + (12.0*thisAngle))));
+    s->position[0] = 250.0f * cf * cosf(11.0f * (thisPointInRadians + 3.0f * thisAngle));
+    s->position[1] = 250.0f * cf * sinf(12.0f * (thisPointInRadians + 4.0f * thisAngle));
+    s->position[2] = 250.0f * cosf(23.0f * (thisPointInRadians + 12.0f * thisAngle));
     
-    rotation = thisAngle*0.501 + 5.01 * (double) s->mystery / (double) BIGMYSTERY;
-    cr = cos(rotation);
-    sr = sin(rotation);
+    rotation = thisAngle * 0.501f + 5.01f * (float)s->mystery / (float)BIGMYSTERY;
+    cr = cosf(rotation);
+    sr = sinf(rotation);
     tmpX1 = s->position[0] * cr - s->position[1] * sr;
     tmpY1 = s->position[1] * cr + s->position[0] * sr;
     tmpZ1 = s->position[2];
@@ -93,14 +95,14 @@ void UpdateStar(global_info_t *global, flurry_info_t *flurry, Star *s)
     tmpY3 = tmpY2 * cr - tmpZ2 * sr;
     tmpZ3 = tmpZ2 * cr + tmpY2 * sr + seraphDistance;
     
-    rotation = thisAngle*2.501 + 85.01 * (double) s->mystery / (double) BIGMYSTERY;
-    cr = cos(rotation);
-    sr = sin(rotation);
+    rotation = thisAngle * 2.501f + 85.01f * (float)s->mystery / (float)BIGMYSTERY;
+    cr = cosf(rotation);
+    sr = sinf(rotation);
     tmpX4 = tmpX3 * cr - tmpY3 * sr;
     tmpY4 = tmpY3 * cr + tmpX3 * sr;
     tmpZ4 = tmpZ3;
     
-    s->position[0] = (float) tmpX4;
-    s->position[1] = (float) tmpY4;
-    s->position[2] = (float) tmpZ4;
+    s->position[0] = tmpX4;
+    s->position[1] = tmpY4;
+    s->position[2] = tmpZ4;
 }
